Foram adicionaod um barra para verficar a segurança na nova senha do usuario no editar perfil,removido comentarios de diversos arquivos do projeto e corrigido um retorno 409 no backend para que o usuario nao edit colocando um nome de usuario ja utilizado

+++ backend/src/user/user.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, NotFoundException,ConflictException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from './user.entity';
@@ -9,8 +9,8 @@ import * as path from 'path';
 
 @Injectable()
 export class UsersService {
-  private readonly saltRounds = 10; // Define o número de rounds para bcrypt
-  private readonly defaultProfileImage = '1.png'; // Nome do arquivo de imagem padrão
+  private readonly saltRounds = 10;
+  private readonly defaultProfileImage = '1.png';
 
   constructor(
     @InjectRepository(User)
@@ -48,36 +48,39 @@ export class UsersService {
 
     return await this.userRepository.save(newUser);
   }
-
   async update(id: number, updateUserDto: Partial<CreateUserDto>, newProfileImage?: Express.Multer.File): Promise<User> {
     const user = await this.findById(id);
   
-    // Se a senha foi atualizada, a senha deve ser criptografada
+    const existingUser = await this.userRepository.findOne({ where: { username: updateUserDto.username } });
+    
+    if (existingUser && existingUser.id !== id) {
+        throw new ConflictException('Nome de usuário já está em uso');
+    }
     if (updateUserDto.password) {
       updateUserDto.password = await bcrypt.hash(updateUserDto.password, this.saltRounds);
     }
   
-    // Se uma nova imagem de perfil foi enviada, delete a imagem antiga
+   
     if (newProfileImage) {
       const imageName = path.basename(user.profileImageUrl);
       if (imageName !== this.defaultProfileImage) {
         this.deleteOldProfileImage(imageName);
       }
-      // Atualize o URL da imagem de perfil no banco de dados
+    
       updateUserDto.profileImageUrl = `http://localhost:4000/users/avatars/${newProfileImage.filename}`;
     } else {
-      // Caso contrário, mantenha a imagem de perfil existente ou a padrão
+     
       updateUserDto.profileImageUrl = user.profileImageUrl || `http://localhost:4000/users/avatars/${this.defaultProfileImage}`;
     }
   
-    // Atualize o usuário com os novos dados
     this.userRepository.merge(user, updateUserDto);
     return await this.userRepository.save(user);
   }
 
+
   async delete(id: number): Promise<void> {
     const user = await this.findById(id);
-    this.deleteOldProfileImage(path.basename(user.profileImageUrl)); // Deleta a imagem de perfil ao remover o usuário
+    this.deleteOldProfileImage(path.basename(user.profileImageUrl)); 
     await this.userRepository.remove(user);
   }
 
@@ -91,11 +94,9 @@ export class UsersService {
     return null;
   }
 
-  // Novo método usando createQueryBuilder
   async findUsersByCriteria(criteria: any): Promise<User[]> {
     const queryBuilder = this.userRepository.createQueryBuilder('user');
     
-    // Adicione condições de busca com base nos critérios fornecidos
     if (criteria.email) {
       queryBuilder.andWhere('user.email = :email', { email: criteria.email });
     }
@@ -104,7 +105,6 @@ export class UsersService {
       queryBuilder.andWhere('user.username IN (:...usernames)', { usernames: usernamesArray });
     }
     
-    // Execute a consulta e retorne os resultados
     return await queryBuilder.getMany();
   }
